Fixes to support 32-bits ASN (macro ASN_SIZE_32) This includes - updates from uint16_t to asn_t - changes in the management of BGP domains (bgp_domain_t) - some functions disabled in bgp/aslevel : topology filters and connectedness/cycle checks

+++ src/cli/bgp.c
@@ -145,7 +145,7 @@ int cli_ctx_create_bgp_domain(cli_ctx_t * ctx, cli_cmd_t * cmd, void ** item)
   }
 
   /* Check that the BGP domain exists */
-  if (!exists_bgp_domain((uint16_t) asn)) {
+  if (!exists_bgp_domain(asn)) {
     cli_set_user_error(cli_get(), "domain \"%u\" does not exist.", asn);
     return CLI_ERROR_CTX_CREATE;
   }
